feat: support patch in VersionInfo
NOTE: This commit introduces a slight alteration to the result of `VersionInfo.paths` for a case where there is no previous version or it's type has changed significantly and includeObjects is true: the root object is also included. This makes sense and fixes an oversight of not including it in the first place.
Signed-off-by: Samppa Saarela <samppa.saarela@iki.fi>