feat: better support for patching (#142)
Old Diff was designed primarily for change detection but used also internally for patching. While path/value "Changes" can be sent over line and applied, they are not very good for that (i.e. result in unnecessary large payloads). While deletions from the head of an array, with this new DiffNode, still result in a large diff/patch, this feature addresses how new object/array values are handled: `DiffNode.patch` contains the new value as-is in case the old value is missing (or it's type has changed). Also deleted object values will not be "exploaded" into path/value -deletions. For example
```ts
const a = {};
const b = { nested: { foo: 'bar' } }

new DiffNode({ oldValue: a, newValue: b }).patch
// => [ { path: '$.nested', value: { foo: 'bar' } } ]

new DiffNode({ oldValue: b, newValue: a }).patch
// => [ { path: '$.nested' } ]
```

`DiffNode.patch` can be used to patch objects using `Path.set` just like the results of the `Diff.changeset`.

NOTE: This update introduces a slight alteration to the result of `VersionInfo.paths` for a case where there is no previous version or it's type has changed significantly and includeObjects is true: the root object is also included. This makes sense and fixes an oversight of not including it in the first place.